既存のTopViewをAccessTokenInputViewに変更して、新しくTopViewを実装した refs #19
