+++ github_api_sample_app_ios_viper/Scenes/Top/TopRouter.swift
@@ -1,30 +1,12 @@
 import Foundation
 import UIKit
 
-protocol TopRouter {
-    func gotoTop()
-    func gotoGitHubUserList(accessToken: String)
-}
+protocol TopRouter {}
 
 class TopRouterImpl: TopRouter {
     let viewController: TopViewController
 
     init(view: TopViewController) {
         viewController = view
-    }
-
-    func gotoTop() {
-        guard let nextView = R.storyboard.top.instantiateInitialViewController() else { return }
-        let configurator = TopConfiguratorImpl()
-        configurator.configure(dependency: nextView)
-        viewController.navigationController?.pushViewController(nextView, animated: true)
-    }
-
-    func gotoGitHubUserList(accessToken: String) {
-        guard let nextView = R.storyboard.gitHubUserList.instantiateInitialViewController() else { return }
-        let configurator = GitHubUserListConfiguratorImpl()
-        configurator.configure(dependency: nextView)
-        nextView.accessToken = accessToken
-        viewController.navigationController?.pushViewController(nextView, animated: true)
     }
 }
